Add modified bootstrap theme from gnome.org/outreachy
There were several different complicated things in this commit.

The easy step was to add theme files (css, js, header and footer images,
minus the sponsor logos) from
https://git.gnome.org/browse/gnome-web-www/tree/theme/outreachy

Next, I added the bootstrap-specific css stylesheet references and
javascript tags to base.html, and duplicated the header and footer from
the gnome.org/outreachy page source.

The tricky bit was how to grab the pages I wanted to link to in the page
header. Each page has a checkbox under the 'promote' tab in wagtail that
sets the show_in_menu boolean flag.

http://docs.wagtail.io/en/v1.11.1/reference/pages/queryset_reference.html?highlight=show_in_menu#wagtail.wagtailcore.query.PageQuerySet.in_menu

I needed to be able to find all live pages in the site that had that
flag set, from the base.html template. In order to set variables in a
django page, I used a context processor:

https://docs.djangoproject.com/en/1.11/topics/templates/#context-processors
https://docs.djangoproject.com/en/1.11/ref/templates/api/#writing-your-own-context-processors

I used those pages, combined with wagtail documentation, to create
home/context_processors.py.

The next step was to set up django to point to the headerpages function
in home/context_processors.py. To do that, I added it to
outreachyhome/settings/base.py, following directions at

https://docs.djangoproject.com/en/1.11/topics/templates/#django.template.backends.django.DjangoTemplates

Once I had a context that was called headerpages, which contained all
the pages with show_in_menu set, I used a for loop to add the pages to
the navigation bar in outreachyhome/templates/base.html.

Signed-off-by: Sarah Sharp <sharp@otter.technology>