Remove references to img/glyphicons-halflings.png
AFAICT by looking at the network console for reloading the outreachy.org
homepage, glyphyicons-halfings.png is never loaded. The image isn't in
the outreachy gnome homepage repo either. The reason this wasn't caught
before is that the browsers never requested it. However, since we're
trying to add persistent storage in dokku, it goes through the css and
makes unique file names for every file referenced in the css. dokku
threw an error when rebuilding the site because glyphicons-halfings.png
was missing.

Signed-off-by: Sarah Sharp <sharp@otter.technology>