Make a pre-scaled logo Streamfield.
This allows us to use the ease of the image streamfield, but make sure
all images aren't bigger than 200x200 pixels.

This is mostly a hack until I can get private sponsor pages internally,
and work out how to pull the sponsor logos out, and display them
dynamically on the home page as we confirm sponsorship.

Signed-off-by: Sarah Sharp <sharp@otter.technology>