Add images and styling to cohort page.
For now, just use the 200x200 avatar images. Gravitar isn't supported
yet. Add bootstrap styling to the text around the images so that they're
left-aligned and text is centered in the middle of the image. Add space
between intern information and mentor/project information. Make intern
names be a smaller heading.