Remove mentor invitation wagtail pages.
Wagtail isn't the right tool for creating community and mentor project
pages. The interface is confusing 'Save draft' and 'Submit for
moderation'. Additionally, the notifications for pages submitted for
moderation doesn't make sense; wagtail notifies a group of page
moderators, but we only want the community's moderator to be notified.

So, we'll ditch Wagtail for the community and mentor pages, and just
write straight Django. We may end up using a rich text field plugin like
django-ckeditor to make it easier for mentors to make "pretty" project
descriptions.