Split out StreamField type-specific templates.
Rather than checking the `block_type` in the template for the HomePage
model, we can construct the Block instances with a path to a custom
template that renders that type any way we want.

This is a little less boilerplate; it makes the connection between
custom block types and how they're rendered explicit; and it makes the
templates reusable if we want to set up other similar StreamFields for
other page types later.