Add form view and template to edit Communities objects
TODO: The community creation needs to be a bit more complex. There's
some data that Outreachy organizers need to approve the community, but
we'll never need that data later. For example, we need to know what
kinds of companies are involved in the community, to ensure that one
company doesn't have too much sway over the community. We'll collect
that information in a custom form later, and email that off to the
Organizers.