Add community approved/rejected/pending views and templates.
Add buttons to the community read-only page for Outreachy organizers to
accept or reject a community to participate in this round.

Modify the following pages to display information about whether a
community is approved, pending, or rejected:
 * community CFP page
 * community read-only page

Hide the 'Submit a project' button on the community read-only page if
the community has been rejected.

Note that the 'rejected' state can be reused for communities that choose
not to participate in this round (either because they don't have funding
or they don't have mentor bandwidth). Change the wording on the
templates from 'rejected' to something softer to account for this.