Add view that handles communities that won't be participating.
Coordinators need a way to indicate a community isn't participating in
this round. Add a new class to add a new participation for this round,
setting some fields to indicate the community isn't participating. Ask
the coordinator why the community isn't participating.

If a coordinator later decides the community can participate, reset
those fields and gather data about their participation in this round.