Make project slug+participation unique together.
This avoids confusion when both a project slug and a community slug are
given to a view: if we found the project slug (which was, after all,
unique) but it didn't match the community slug, what should that mean?

Now we need both specified to uniquely identify a project, so if we
don't check one or the other then we'll get errors, which hopefully
we'll spot in testing.