Add a project skill model
One of the most common questions I hear from applicants is "how much do
I need to know about X". Mentors often list skill descriptions with no
indication whether an applicant needs to be a beginner or an expert.
Applicants who come from groups under-represented in tech are more
likely to have impostor syndrome and more likely to under estimate
themselves because of systematic discrimination.

Force mentors to be very explicit about the experience level they expect
applicants to have with a particular skill, and whether that skill is a
hard requirement, a mentor preference, or an optional bonus.

Try to ensure that we have a description that fits all types of skills,
from programming languages, translation, documentation, user experience,
and graphical design.