Reorder approval-status choices; default to WITHDRAWN.
For approvers and submitters, the PENDING and APPROVED states are first
and second most important, and getting the model's choices attribute to
reflect that lets us write code that just loops over the choices.

The default status for new objects should be WITHDRAWN because before
being created that is effectively the status they had anyway, so making
it explicit that going to PENDING is a state change makes permission
checks and state-transition actions simpler. This should have no effect
on the current code because everything explicitly sets the status on
creation anyway, but...

As long as Django is forcing me to do a useless migration that has no
effect on the database because of reordering the choices, I might as
well throw in a migration that actually does something I forgot to take
care of before.