Add all types of coordinator sign up workflow.
Fixes #73
Fixes #53

There are four situations to handle for coordinator workflow:
 1. Someone submits a new community
 2. Someone signs up as a coordinator for a community that is not
    participating yet
 3. Someone signs up as a co-coordinator for a community that is
    approved to participate
 4. An approved coordinator signs up a community to participate

Scenario #1 - New community, new coordinator
-----------

Person submits a new community.
A CoordinatorApproval is created with the status set to approved.

The new community creation triggers an email to the Outreachy organizers
to approve both the community participating in this round and the
coordinator who submitted this community.
(community-signup.txt email template)

Once the organizers approve that community, it triggers an email to all
approved coordinators that the community is approved
(community-approved.txt email template)

Scenario #2 - Old community, new coordinator or co-coordinator
-----------

The person picks a community from the list of past participating
communities. They sign up to be a coordinator.
A CoordinatorApproval is created with status set to pending.

The CoordinatorApproval status set to pending triggers an email to the
approved community coordinators and the Outreachy organizers to review
the coordinator.
(coordinator-review.txt email template)

Either a coordinator or an organizer approves the new coordinator.
This triggers an email to the newly approved coordinator
(coordinator-approval.txt email template)

One of the coordinators signs the community up to participate in the
current round. This triggers an email to the organizers for review.
(community-signup.txt email template)

The organizers approve the community, which triggers an email to all
approved coordinators.
(community-approved.txt email template)

Scenario #3 - Participating community, co-coordinator signup
-----------

The person picks a community from the list of past participating
communities. They sign up to be a coordinator.
A CoordinatorApproval is created with status set to pending.

The CoordinatorApproval status set to pending triggers an email to the
approved community coordinators and the Outreachy organizers to review
the coordinator.
(coordinator-review.txt email template)

Either a coordinator or an organizer approves the new coordinator.
This triggers an email to the newly approved coordinator
(coordinator-approval.txt email template)

Scenario #4 - Old community, approved coordinator
-----------

One of the coordinators signs the community up to participate in the
current round. This triggers an email to the organizers for review.
(community-signup.txt email template)

The organizers approve the community, which triggers an email to all
approved coordinators.
(community-approved.txt email template)