Dashboard: Show all coordinators you are approver for.
Fixes #81. Closes #82 (which this patch subsumes).

Many thanks to @gedankenstuecke and @madprime for being best possible
inadvertent beta testers: correctly identifying the root cause of the
problem and for proposing a fix that was almost spot on! ;-)

The only difference from the patch in #82 is that we only show you other
coordinators in your community if you are already approved as a
coordinator. If you have requested to be a coordinator but not been
approved, you'll only see yourself, which is allowed because you also
are a submitter in that case.

This function now closely resembles the corresponding function in
MentorApproval, which is probably a good sign.