Make login form more clear.
New users keep starting to put information into the account fields,
thinking they're signing up, stopping at the "Log in" button, and then
going back to find the "Create one." button.

Most people aren't going to log out of the site, so the majority of
people are going to be creating new accounts. Move that button up top
and make it the same color as the login button.