Implement CoC prompt for existing users who haven't agreed to the CoC.
The agreed_to_code_of_conduct field started out as a BooleanField.
The migration made the value for all existing users be False.

I couldn't figure out how to validate it or force it to be checked.
So it was changed to a CharField (which has blank=False by default).
Now the existing users have the field set to "0".
Check that in the template.
No one should have "0" as their legal name, hopefully...