Simplify final-application URLs.
The primary key of the FinalApplication object isn't needed because
there's a unique_together constraint on project and applicant, so since
we already have both of those in the request we don't need to
disambiguate further.

However we do have to actually do something with the `username` part of
the path, if present, and set it correctly when necessary.