Simplify finding approved mentored projects.
In community_applicants, we've already checked that the Participation is
approved (or we'd return a 404 response), so checking that in
get_approved_mentored_projects was redundant in this case.

We also already picked (the current) RoundPage in the view, and if we
pass the Participation to the template instead of the Community then we
have a reference to that RoundPage.

So given a known-approved Participation, we only have to find the
Projects related to that Participation and limit them to approved
Projects, and both of those filters can be done purely in the template.