Require a Comrade before making contributions.
These two views assumed that user.comrade exists, but there was nothing
to enforce that, so they'd just raise exceptions.

Note that I removed the explicit addition of 'user' to the template
context for eligibility results, because we have the
django.contrib.auth.context_processors.auth context processor turned on,
which already adds `request.user` to the context as `user`.