Redirect applicants who haven't completed an eligibility check from the contributions page.
Add text to the two pages that might be shown (the eligibility results
page for folks who have a pending application or the project
contributions page) to explain a little more about the next steps in the
process.

Link to the rounds page in a way that doesn't detract from the applicant
recording contributions and making their final application.

Add the link to the contributions page to the mentor reminder prompt,
but warn them it only works for approved applicants (not mentors).

Fixes #160