Add a way for mentors to resign.
Sometimes in the middle of the round, a mentor needs to resign from an
internship, and we find a new mentor. In this case, they need to sign
the mentor agreement, but we don't want to remove the intern from the
program.

Thus, we need a way to destroy a MentorRelationship without destroying
an InternSelection, and keeping the past signed mentor agreement around
for legal reasons.