Use fully-qualified names for date/time types.
The name "datetime" can refer to a class in the "datetime" module or to
the module itself. The name "time" can also refer to a class in the
"datetime" module or to an entirely different module in the Python
standard library. In both cases it is more clear to refer to e.g.
"datetime.datetime" or "datetime.time".

Also, models.py uses local variable names including "date" and
"timezone", and shadowing global names like that is likely to lead to
confusion later. Ensuring we only use the global names fully-qualified
avoids most of the possibilities for confusion.