Fix internship payment dates to align with SFC expectations
Only show the latest date, because interns often get anxious and ask
questions of the accounting department if they don't receive their
payment on the very first possible day.

The payout date calculations are a little... odd, but next round we
should have them be an actual field that can be changed.