Honor Dokku-provided X-Forwarded-For header.
Make sure the specified address matches the IP address of the reverse
proxy that gunicorn receives connections from. This is the default value
if you deploy in Docker, and you haven't changed Docker's default
subnet, and the reverse proxy is running on the host machine that's
running Docker. These conditions apply to the default Dokku
installation, for example.