Add automated email to mentors to confirm interns have started.
The Outreachy organizers need to ensure all interns have started their
internships before we can authorize the first $500 stipend to be sent.
Mentors have until a specific date to let us know if their intern hasn't
started.

Add the Outreachy organizer email address to the To header to ensure we
sent the messages.