Delete a redundant duplicate-email warning.
Since the account creation form now reports an error if you try to
create a new account with the same email address as an existing account,
we don't need to warn people ahead of time about this relatively
uncommon (<3% of users) process failure mode. The warning probably just
contributes to people ignoring the text on these pages.