Include missing "hidden" fields in sponsor form
Without these fields we can get an error on form submit that looks like:

django.utils.datastructures.MultiValueDictKeyError: "'sponsorship_set-0-id'"

When editing an inline form that has an object that already exists, the
internal database primary key is included as a hidden field in the form
so that the server can remember that the newly-submitted values should
be saved into the existing object instead of creating a new object.

Fixes #222.