Policy change on initial application rejection
If based on our gender and demographics questions on the initial
application, we're unsure if an applicant further Outreachy's program
goals of supporting people from groups under-represented in the
technology industry, put the application in the pending state, not
rejected state.

An Outreachy organizer will then review the application, and determine
its approval status.