Add new PaymentEligibility class linked to ApplicantApproval
Two fields are moved from ApplicantApproval to PaymentEligibility:

 * us_national_or_permanent_resident
 * living_in_us

They are changed from a NullBooleanField to a BooleanField, so we should
now get validation at that step in the wizard that they have been set to
either 'Yes' or 'No'.

The connotations of living_in_us is changed. Now it asks if the
applicant is living in the U.S.A. for any time during the internship
period (and five weeks after the internship ends). Before it asked if
the intern was living in the U.S.A. for the "majority" of the
internship.

We need to know if the applicant is planning to work from the U.S.A.
because they may run afoul of tax law if Conservancy pays them while
they're in the U.S.A. or pays them for work done while residing in the
U.S.A. and their visa doesn't allow full-time work from the U.S.A.
Conservancy cannot provide tax advice to applicants, but we do need the
information to determine whether the intern should fill out a W8-BEN or
a W-9.

The living_in_us question used to be used to decide whether to display
racial identities questions. If the person was living in the U.S.A., we
would display them. However, that means we would be asking questions
about race for European citizens, which has legal issues.

Now we'll only display questions about race to U.S.A. citizens or
permanent residents. I suppose someone might be a dual citizen with the
U.S.A and a European country? If we find this is an issue, we can ask
about whether someone is a European Union member country citizen.