Add question about whether students need CPT documentation.
Students who are in the U.S. on an F-1 visa have to apply to their
university if they want to participate in an internship. This is called
curricular practical training (CPT). Universities can't let them work
during dates they have classes. Outreachy can accomodate shifting the
internship sooner (up to when the internship is announced) or later (up
to five weeks later). We do have to provide documentation for the
interns to give to their university with the modified dates for their
internship.

Outreachy process for handling CPT:

We don't care about CPT until interns are selected, because we expect
the number of interns who need this documentation to be low. This might
change if we get many applicants who need CPT documentation, or whose
dates turn out not to work, even with shifted internship dates.

When mentors select an intern, we follow up with them via email to
confirm the dates for their CPT will work. We send their mentor a
message to tell them we have to confirm their CPT dates, and they should
have a backup in mind in case this intern isn't eligible.

We could have the CPT eligibility dates as part of the initial
application, but I don't have enough time right now.