Handle empty ApplicantGenderIdentity.
Fixes "IndexError: list index out of range" when saving an application
if the applicant did not choose any gender identities.

Given that "prefer not to say" is one of the options on that page, I
think we should require that _something_ be checked, but I haven't had a
chance to check that idea with Outreachy organizers yet.