Update views and templates for requesting essay revisions and providing them.
Lower my stress by only showing the number of pending essays that I
haven't requested essay revisions to on the staff dashboard.

Display the pending applications that need an essay revision in a
separate (lower) table on the pending application summary page.

On the applicant detail page, add a button to request an essay revision.
This will set the new flag, and email the applicant to tell them they
need to revise their essay.

Only applications with that bit can be edited. Once the essay is
re-submitted, it cannot be edited again. This minimizes the amount of
churn reviewers have to go through with an applicant that is constantly
revising their essay answers because they're anxious.