Bug fix: Same objects with different querysets can't be compared with 'in' keyword.
We think that if the same object is found with a different queryset, the
pointer to the object is different. Therefore, if you try to test if the
object is in a list, it will return false, even if the object is in the
list, because the pointer values are different.

Let's test this hypothesis.