Add date picker automatically to all DateFields.
Now whenever a DateField is displayed in a form, it will automatically
become a date picker, and be shown in the browser's locale-specific date
format. This should help with everyone who struggled with putting dates
in YYYY-MM-DD format. Hopefully we get less people switching the years
they are going to school in, or picking the wrong months.

Update all field verbose names and help text to remove references to the
YYYY-MM-DD format.