Remove old email for initial payment reminder
Before we switched to using three payments, we automatically sent out
the initial payment. Now we ask mentors to give feedback on their
interns after the first week, to ensure they've started working. Remove
the email that referenced automatic initial stipend payments, and the
link from the staff dashboard to send that email.