Project selection page now handles having no current round.
Once the application period is closed, it doesn't make sense to show
internship projects any more. Instead, we show a generic timeline and
link to past rounds. We show a message about intern selections being
available on the alums page after the intern announcement date. We
show the section about how to prep for the application period. If
there's a previous round, we display the common skills from the previous
round.

Write tests that show the applicant's views into the project selection
page are correct.