Fix bug with initial feedback revision tracking.
When a mentor submits or edits an initial feedback, we record their IP
address in the InitialMentorFeedback class. The class is under revision
control with the django-revision package. However, staff members must
approve the initial payment, and that field is in the
InitialMentorFeedback class. That means the latest revision may be a
staff member changing the payment status, rather than a mentor
submitting the feedback. We have the further problem that changes to the
parent class, InternSelection, by staff can also cause a revision to be
created.

For now, search for a revision that was made by an approved mentor for
the intern. This won't cover the rare case where we need to find a new
mentor in the middle of the round. The mentor would need to not withdraw
from the project in order for this code to be able to dig out older
revisions.