Travel stipend: Add active rounds.
All the rounds with active travel stipends are stored in RoundPage
objects on the Outreachy website. Add a new method to the RoundPage
class to ask if the travel stipend is active, and when the deadline is
for starting travel.

If we ever decide to fill in the history of round pages by adding
CohortPage objects, the CohortPage will need to have a field added to it
for when the interns were announced. For now, manually list all older
rounds with inactive travel stipends in the template.