Remove names of Outreachy PLC
In order to keep the Outreachy email address from being found by bots,
we need to allow people to contact us through the web form. That means
we can't currently use the Outreachy PLC email, because we only have the
Outreachy organizers email set up as a form.

The Outreachy organizers can forward travel requests to the Outreachy
PLC or Karen, as necessary.