Make admin email views take a round slug
This centralizes responsibility for deciding which round to send email
for into the dashboard. The individual email views will construct emails
for any round you ask them to.

In the process, I made the dashboard do the check for a currently-active
round that send_biweekly_internship_email was doing, which means that
the dashboard won't display links where previously the view would have
refused to generate any email.