Fix CFP sorting when the CFP is not open
Also, when the CFP _is_ open, all communities that exist in the database
need to be displayed somewhere, because otherwise people try to add them
again and that triggers a database integrity error. If some community is
spam and should not be displayed at all, we'll just delete it using the
Django admin.