Use correct round for contributor emails
When the {% timezone %} tag is used in a template, it doesn't matter
what timezone any datetime objects are in; they'll get converted to the
specified one. So these model methods didn't need access to
comrade.timezone, but they did need access to the right round, so move
them over there.