Remove unused applicantapproval-pending.txt template
1. This template was never used because applicant_approval_status_change
was only called from ApplicantApprovalUpdate, which is only used by
approvers, who can't put an object into pending state.

2. This template could not be used because it had a syntax error.

3. If this template were used, it would go to the Outreachy organizers
instead of to the applicant, which is contrary to who the text appears
to be intended for.

Having a template by this name was sort of required at the time because
approval_status_changed would have blindly tried to use it if it was
called with a pending applicant approval object. But now that case just
results in silently not sending email, so it's fine to delete this.