Simplify snippet logic in round_page_with_communities.html
Note that if current_round is not None, then due to the conditions in
the view, has_late_application_deadline_passed is False and can be
removed without changing behavior.

The rest is just some reordering of checks to make them easier to reason
about.

There shouldn't be any functional changes in this commit.