Add Comrade.approved_volunteer helper
All but one use of approved_mentor_or_coordinator also checked
approved_reviewer, so all those cases are simplified by checking a
single broader condition.

The remaining case was just determining whether to display an
eligibility prompt, and encouraging an application reviewer to put in
their own application, um, doesn't make sense. So it should have
included this check too.

I'm calling the broader class of people "volunteers" even though some of
them may be acting in capacities which they're paid for, because Sage
and I can't come up with a better simple term for this.