Add custom mid-point feedback dates per intern.
Add two new fields to InternSelection that say when the mid-point
feedback forms will open, and when mid-point feedback is due.

If an intern has a custom initial feedback date, calculate the mid-point
feedback dates as six weeks after that custom date. (I debated whether
to back-calculate from the intern's custom end date, but those should be
the same.)