manage.py test doesn't need a what-to-test argument
I think I remember that historically, manage.py test would run the tests
in other apps used by the current project, which would be annoying
because presumably those apps ran their own tests before release. But if
that was ever true, it isn't now. Hooray!

However, I do want to turn the verbosity way up on the test runner when
running on Travis-CI, because it's hard to investigate what went wrong
when the tests ran in a CI environment.