Deadline has passed on the deadline date too
We've reached a given (4PM UTC) deadline once it is _equal to_ "now", so
`has_deadline_passed` needs to check if the deadline is less than _or
equal_ to today.

Before I broke this in abd08668d0b0892d956bc287389c14dc404df089, this
function was comparing datetimes with sub-second resolution, so missing
the equality check was only wrong for a fraction of a second. Now that
the resolution is 24 hours instead, this is wrong for... 24 hours. And
that's too many hours.