Change approved_to_see_all_project_details to use new Role class.
There was also redundant code.

if Participation.participating_round.is_mentor(user):

is a superset of

if Participation.is_mentor(user):

In the is_mentor function for the Participation class, we see if this
Comrade is mentoring an approved project for this community in this
round (the Participation):

    def is_mentor(self, user):
        try:
            return user.comrade.get_mentored_projects().approved().filter(
                project_round=self,
            ).exists()
        except Comrade.DoesNotExist:
            return False

In the is_mentor function for the RoundPage, we see if this Comrade is
mentoring an approved project for any community in this round:

    def is_mentor(self, user):
        try:
            return user.comrade.get_mentored_projects().approved().filter(
                project_round__participating_round=self,
                project_round__approval_status=ApprovalStatus.APPROVED,
            ).exists()
        except Comrade.DoesNotExist:
            return False

If we're checking for if either condition is true, than we only need to
call Participation.participating_round.is_mentor(user). That call is
equivalent to Role.is_mentor():

    def is_mentor(self):
        if self.current_round is not None and self.user.is_authenticated:
            return self.current_round.is_mentor(self.user)
        return False