Fix mentor approval confirmation page.
Before, the page would display the mentor profile information. The
button to confirm approval of the mentor would be at the very bottom of
the page. That meant many coordinators would not click it.

Make the information on the page much shorter if a coordinator is
approving or rejecting a mentor. All the mentor profile information can
be found on the project page that we link coordinators to in the email.