Ask applicants what their job title and job duties are.
Display a warning on the applicant review if they have a part-time job.
This displays regardless of the number of hours, because it's too hard
to compute that in a template.

Display their job title and job description to applicant reviewers (this
may also display it to mentors, I don't know?).

Fixes #272