Rework student term collection and eligibility rules around hemisphere.
Add a new field last_term to say that this SchoolTimeCommitment is the
last term in someone's degree.

Require all students to provide their current term and two terms after
that, except if they're graduating and one of the terms is their last
term.

Verify:
 - terms are in consecutive order
 - only one term has last_term set
 - all three terms are filled out, unless one term has the last_term set

Update the eligibility rules and eligibility results page to say that
students are restricted to specific Outreachy rounds based on which
hemisphere their school is located in. Reference the Wikipedia page on
hemispheres in case they don't know what the term means.

Notify the students that while they can upload their school calendar to
a file sharing service, the Outreachy organizers may not take into
account any calendar except a publicly available official school
calendar.