Fix a bug when co-mentors sign up after payment dates change
This fixes a IntegrityError that occurs when an intern was selected, the
payment dates (and feedback dates) changed, and then a co-mentor tried
to sign up.

The problem is that the get_or_create() function will look for an
InternSelection with the new feedback dates and the InternSelection with
old feedback dates won't match. Then it will try to create a new
InternSelection, and the uniqueness constraints on that class (the
combination of ApplicantApproval and Project must be unique) will not be
met.

Fix this by looking for an InternSelection that matches the
ApplicantApproval and Project only, and then just use the create()
function if we can't find it.