Change name of svg and increase size.
Make the svg reflect the CamelCase names of the classes, because it's
easier to copy and paste (and matches searches for the class name).

GitHub scales the image in the documentation to 555 pixels wide on a
desktop browser. Let's export the png to be 500 pixels wide and see how
it looks.