Mentor dashboard: Display intern selection results
During the intern selection process, display a warning box for mentors
to keep their intern selections private until the announcement date. The
box will show if the Outreachy organizers have approved their intern, if
the intern selection is still pending, or if the intern was not
accepted.

If the intern was not accepted, don't show the details of the internship
round schedule for payments and feedback.