Send mentor signature requests exactly once per intern+mentor
If we've just approved an additional MentorApproval for a project which
already has one or more InternSelections, we only need to notify the new
mentor about those interns---and fortunately, we already know who the
new mentor is, because we just added them.

By contrast, when adding a new InternSelection, we need to notify all
mentors for the selected intern's project---except for the person who
just selected that intern, because in the process they signed the mentor
agreement, so we don't need to ask them to do it again.