Improve description of Outreachy blog post chats in prompt emails.
The goal of these changes is to ensure that both Outreachy mentors and
interns participate in the bi-weekly chats. We want to have the chat at
a specific time, to encourage interns to have a deadline for thinking
about their blog post topic. Having the chat at a specific time will
also encourage Outreachy interns to interact with each other.