Update initial feedback instructions email
Remove field asking for mentors to submit feedback to be sent to the
mentors' mailing list. Only 14 out of 46 mentors did so last round. I
don't have time to send a summary email to the mentors' mailing list.

We also didn't include the option for the mid-point or final feedback,
so make the initial feedback consistent with the other feedback emails.

Don't say interns will receive an email when their payment authorization
is sent out to Conservancy. We didn't implement that feature. Instead,
direct them to look at their dashboard, which shows the payment
authorization has been sent. This makes the instructions consistent with
the midpoint and final feedback instructions.