Fix test case failing that fails after 00:00 UTC
Don't override the start date when calling into the round factories.
There's magic code to pick a start date that will allow the deadline to
be past when the timezone is in UTC. This test only succeeded because
the code was written before 5 pm PST or 0:00 UTC.

Code says: You caught a bug because you're working late. Stop working
late!